make quoted capture less greedy when we have unambiguous separator
In [logstash-plugins/logstash-filter-kv#60][], GitHub user @robcowart reports
that in some scenarios we can fail to split on match of `field_split_pattern`.

The example given is a partially-quoted value, with additional bytes between
it and the unambiguous separator, followed by another key and quoted value.

Since the quoted value isn't immediately followed by a field splitter, our
semi-greedy quoted-value capture was too greedy, continuing to consume until
it found a close-quote that would be followed by either a field-split or EOF.

Here, we become much less greedy, capturing any sequence of characters that
is _not_ the close-quote character.

[logstash-plugins/logstash-filter-kv#60]: https://github.com/logstash-plugins/logstash-filter-kv/issues/60

Fixes #62