adds `whitespace => strict` mode
When configured with a space for its `field_split` (as is default), the
leniency in acceptance of optional spaces surrounding the `value_split`
prevents the parser from being able to detect an unquoted "empty" field.

By adding an opt-in whitespace mode that is strict and does not allow
unnecessary whitespace, we enable users who behave in this strict manner
to specify "null" or "empty" fields in their events.

Resolves: #65
Resolves: #37