handle runtime errors instead of allowing plugin to crash
While this plugin is designed to be flexible with inputs, there are situations
where it calls out to external APIs (such as the Logstash Event API) that can
raise runtime exceptions.

When an exception is raised, tag the event with `_kv_filter_error` and warn
instead of allowing the exception to blow through the stack and crash the
plugin. If the logger has debug enabled, include the exception's backtrace.