escapes: support `field_split_pattern` beginning with literal backslash
Previously the `unquoted_capture_until_pattern` helper would _always_ capture
a backslash and the following character, assuming it to be an escape sequence;
when presented with a `field_split_pattern` that begins with a literal
backslash, this caused the beginning of the field splitter to be captured.

By moving the backslash-any matcher to after the negative lookahead, we can
ensure that we break the capture when encountering the field split pattern,
while continuing to capture backslash-escape sequences.